pkg/api: fix singular name
With the 1.27 Kubernetes upgrade, a new GetSingularName method needed to
be implemented by the metrics rest.Storage. While doing it, we set a
value for the singular name of the resources when it used to be empty.
Due to that, the kube-apiserver stop guessing the resource name based on
its Kind, which broken kubectl get PodMetrics/NodeMetrics.

To fix that, we need to blank out the singular name of the resources,
otherwise we will be stuck with the following commands:
- kubectl get pod.metric
- kubectl get node.metric
Instead of the more commonly used:
- kubectl get PodMetrics
- kubectl get NodeMetrics

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>
(cherry picked from commit a9fdc887cfd06d2991940ca15e454f77261e678f)